[BE] 모임 멤버 탭 관련 기능 추가

+++ backend/src/test/java/moim_today/implement/meeting/joined_meeting/JoinedMeetingAppenderTest.java
@@ -69,7 +69,7 @@ class JoinedMeetingAppenderTest extends ImplementTest {
         // given 2
         JoinedMoimJpaEntity joinedMoimJpaEntity = JoinedMoimJpaEntity.builder()
                 .moimId(moimJpaEntity.getId())
-                .memberId(MEMBER_ID.intValue())
+                .memberId(MEMBER_ID.longValue())
                 .build();
 
         joinedMoimRepository.save(joinedMoimJpaEntity);
